# 8333154: RISC-V: Add support for primitive array C1 clone intrinsic